Add a.vim
Note: a.vim defines shortcuts with leader prefix and multiple keys in
insert mode. This causes a short pause when typing leader (<Space> in
my conf) in insert mode. So I had to remove these mappings to remove the
insertion pause. Since the plugin mappings are loaded after the vimrc
file, I had to unmap them at the VimEnter event.